Add basic filtering section Reason: You can now do additional filtering by adding or removing an advanced query lookup value. - Ability to exclude options not yet implemented. - Ability to load more field options not yet implemented. - Price field is not yet implemented. - Search Options field not yet implemented. - Heavy styling still required.
How this change addresses the issue:
A new reducer is made to hold a list of strings, each pertaining
to a field-option combination, ex: "(@tppays=="France")"
or "(NOT @tpcategorie=="Vin rouge")"
When an option is checked or unchecked, the actions
'updateAdvancedQuery' or 'removeAdvancedQuery' are called.
These actions add/update (i.e. modify the 'exclude' value) or remove
a field-option combination.
When the list of field-option combination strings are modified,
the useEffect func in SearchMain will trigger again, fetching using
the new set of advanced queries.