Add URL and Redux syncing Reason: Accessing the site with URL queries will now set the initial Redux state or modify the existing Redux state accordingly. Likewise, modifying the Redux state will affect the URL. This feature is done by utilizing a 'HistoryState' component which is strictly used for setting the URL to the Redux state or vice versa, by monitoring the state and url and converting the data using two useEffect hooks.
How this change addresses the issue:
Query results can now be linked/shared to other users.

Side Effects:
Currently, neither the Redux store or the URL is the source of truth.
When the page is first loaded, and when the back/forward browser buttons
are pressed, the URL adjusts the store. In all other cases, the Redux
state adjusts the URL.
There are two other options that I think would have been much simpler:
1) The URL is the source of truth, and no duplicate data would be in the
   store.
2) The Redux store is the source of truth. The only time the URL is
   synced to the store is when the App first loads, at which point the
   URL params will be cleared. Additional, a 'Share This Page' button
   will be made available to grab a link to the current query.
I only went with the method of syncing url and state to practice, but I
think it isn't the simplest nor the most maintainable way, unfortunately